mirror: add --dry-run

+++ src/mirror.py
@@ -16,12 +16,13 @@ PROJECT_NAME = 'mirror'
 CONF = cfg.CONF
 opts = [
     cfg.BoolOpt('debug', help='Enable debug logging', default=False),
-    cfg.StrOpt('images', help='Path to the images.yml file', default='etc/images.yml'),
-    cfg.StrOpt('server', help='SFTP server'),
+    cfg.BoolOpt('dry-run', help='Do not really do anything', default=False),
     cfg.IntOpt('port', help='SFTP port', default=22),
-    cfg.StrOpt('username', help='SFTP username'),
+    cfg.StrOpt('basepath', help='Basepath', default='/'),
+    cfg.StrOpt('images', help='Path to the images.yml file', default='etc/images.yml'),
     cfg.StrOpt('password', help='SFTP password'),
-    cfg.StrOpt('basepath', help='Basepath', default='/')
+    cfg.StrOpt('server', help='SFTP server'),
+    cfg.StrOpt('username', help='SFTP username')
 ]
 CONF.register_cli_opts(opts)
 CONF(sys.argv[1:], project=PROJECT_NAME)
@@ -65,22 +66,23 @@ for image in images:
         except OSError:
             logging.info("'%s' not yet available in '%s'" % (filename, dirname))
 
-            logging.info("Downloading '%s'" % version['source'])
-            response = requests.get(version['source'], stream=True)
-            with open(os.path.basename(path.path), 'wb') as fp:
-                shutil.copyfileobj(response.raw, fp)
-            del response
+            if not CONF.dry_run:
+                logging.info("Downloading '%s'" % version['source'])
+                response = requests.get(version['source'], stream=True)
+                with open(os.path.basename(path.path), 'wb') as fp:
+                    shutil.copyfileobj(response.raw, fp)
+                del response
 
-            if fileextension in ['.bz2', '.zip']:
-                logging.info("Decompressing '%s'" % os.path.basename(path.path))
-                patoolib.extract_archive(os.path.basename(path.path), outdir='.')
-                os.remove(os.path.basename(path.path))
+                if fileextension in ['.bz2', '.zip']:
+                    logging.info("Decompressing '%s'" % os.path.basename(path.path))
+                    patoolib.extract_archive(os.path.basename(path.path), outdir='.')
+                    os.remove(os.path.basename(path.path))
 
-            logging.info("Uploading '%s' to '%s'" % (filename, dirname))
-            try:
-                client.mkdir(dirname)
-            except Exception:
-                pass
+                logging.info("Uploading '%s' to '%s'" % (filename, dirname))
+                try:
+                    client.mkdir(dirname)
+                except Exception:
+                    pass
 
-            client.put(filename, os.path.join(dirname, filename))
-            os.remove(filename)
+                client.put(filename, os.path.join(dirname, filename))
+                os.remove(filename)
